+++ modules/hdf5/src/cpp/h5ReadFloat.cpp
@@ -65,13 +65,15 @@ h5ReadFloat(hid_t dset_id, hid_t type_id, hid_t dspace_id, bool asAttribute, std
             hsize_t* h5_dims = nullptr;
             hsize_t* h5_maxdims = nullptr;
             try {
-                h5_dims = (hsize_t*)new_with_exception<hsize_t>(rank * sizeof(hsize_t), false);
+                h5_dims
+                    = (hsize_t*)new_with_exception<hsize_t>((size_t)rank * sizeof(hsize_t), false);
             } catch (Exception& e) {
                 error = e.getMessage();
                 return ArrayOf();
             }
             try {
-                h5_maxdims = (hsize_t*)new_with_exception<hsize_t>(rank * sizeof(hsize_t), false);
+                h5_maxdims
+                    = (hsize_t*)new_with_exception<hsize_t>((size_t)rank * sizeof(hsize_t), false);
             } catch (Exception& e) {
                 error = e.getMessage();
                 return ArrayOf();
