+++ modules/hdf5/src/cpp/h5ReadVlen.cpp
@@ -191,13 +191,14 @@ h5ReadVlen(hid_t attr_id, hid_t type, hid_t aspace, bool asAttribute, std::wstri
         hsize_t* h5_dims = nullptr;
         hsize_t* h5_maxdims = nullptr;
         try {
-            h5_dims = (hsize_t*)new_with_exception<hsize_t>(rank * sizeof(hsize_t), false);
+            h5_dims = (hsize_t*)new_with_exception<hsize_t>((size_t)rank * sizeof(hsize_t), false);
         } catch (Exception& e) {
             error = e.getMessage();
             return ArrayOf();
         }
         try {
-            h5_maxdims = (hsize_t*)new_with_exception<hsize_t>(rank * sizeof(hsize_t), false);
+            h5_maxdims
+                = (hsize_t*)new_with_exception<hsize_t>((size_t)rank * sizeof(hsize_t), false);
         } catch (Exception& e) {
             error = e.getMessage();
             return ArrayOf();
